webkit/data/layout_tests/platform/chromium-win/LayoutTests/http/tests/security/XFrameOptions/x-frame-options-parent-same-origin-allow-expected.txt

http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-parent-same-origin-allow.cgi - willSendRequest <WebRequest URL "http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-parent-same-origin-allow.cgi">
<unknown> - didFinishLoading
There should be content in the iframe below



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: This should show up as the parent is in the same origin.
